// apps/restaurant-web/src/types/api.ts
export interface MenuCategory {
  id: number;
  name: string;
  icon: string;
  imageUrl?: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  categoryId: number;
  imageUrl?: string;
  branchId?: string | null;
}